+++ src/Slices/Notification/Core/Query.ts
@@ -1,4 +1,5 @@
 import { PageSize, Pagination } from "@/Core/Domain";
+import { RemoteData } from "@/Core/Language";
 import { Model, Severity } from "./Model";
 import { Origin } from "./Utils";
 
@@ -43,3 +44,13 @@ export const drawerQuery: Query = {
   pageSize: PageSize.from("100"),
   filter: { cleared: false },
 };
+
+export type ViewData = RemoteData.RemoteData<
+  Manifest["error"],
+  Manifest["usedData"]
+>;
+
+export type ApiData = RemoteData.RemoteData<
+  Manifest["error"],
+  Manifest["apiResponse"]
+>;
